requirements.txt
numpy==1.26.4fuzzywuzzy==0.18.0ollama==0.2.1opencv-python==4.6.0.66opencv-python-headless==4.7.0.72openai==1.37.0pandas==2.2.2pdfplumber==0.7.6PyMuPDF==1.22.1pytesseract==0.3.10python-box==7.1.1python-dateutil==2.9.0.post0PyYAML==6.0.1python-dotenvuvicornfastapianthropicpython-multipart